requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
ta==0.11.0
